Duplicated the main scene for a second level
made a second level (duplicated the current scene for it). Also made and ending scene and started on the credits scene (making the credits to put in it). Ending scene will link back to the main menu scene but you can't currently get to the second level or the ending scene.