Credits and End scene mostly complete with second level underway
I have included some of the credits into the credits scene at the start. The end scene now displays a screen saying "Game Over!" when the country health reaches zero. Currently if you hit a switch it takes you straight to level 2 and then if you hit a switch in level 2 it will restart level 2 which is a bug that will be fixed very soon. There is now a platform for the player to start on which is above the ground and I have made the country health in the top left corner slightly transparent so that if the player goes into it you can still see the character through the health bar.