Minor Fixes and added in play again to level 1 to the losing scene and win scene
Some minor bug fixes including so text sizing for the level 2 timers. I also added to the how to play scene as I had forgotten to put in the W, A and D keys. I also added a main menu button that takes you directly to the main menu for the game over scene and also the you won scene. In addition to that I changed the play again button from taking you to the main menu and now when clicked it takes you straight to the first level. This was requested by family who tested the game out. Another bug where the player was regestering as moving to the right when the key was pressed but was endlessly changing between the idle animation and the run animation was fixed by changing it from an if statement to an else if statement. I think this is because I updated my visual studio and they changed some of the conditions for if and else if statements.